+++ src/codemagic/cli/cli_process.py
@@ -7,6 +7,7 @@ import subprocess
 import sys
 import time
 from typing import IO
+from typing import Dict
 from typing import Optional
 from typing import Sequence
 from typing import Union
@@ -79,14 +80,22 @@ class CliProcess:
         if self._process.stderr:
             self._stderr_stream = CliProcessStream.create(self._process.stderr, sys.stderr, blocking=False)
 
-    def execute(self,
-                stdout: Union[int, IO] = subprocess.PIPE,
-                stderr: Union[int, IO] = subprocess.PIPE) -> CliProcess:
+    def execute(
+            self,
+            stdout: Union[int, IO] = subprocess.PIPE,
+            stderr: Union[int, IO] = subprocess.PIPE,
+            env: Optional[Dict[str, str]] = None,
+    ) -> CliProcess:
         self._log_exec_started()
         start = time.time()
         try:
             if not self._dry_run:
-                self._process = subprocess.Popen(self._command_args, stdout=stdout, stderr=stderr)
+                self._process = subprocess.Popen(
+                    self._command_args,
+                    stdout=stdout,
+                    stderr=stderr,
+                    env=env,
+                )
                 self._configure_process_streams()
                 while self._process.poll() is None:
                     self._handle_streams(self._buffer_size)