+++ src/codemagic/shell_tools/__init__.py
@@ -1,0 +1,1 @@
+from .keytool import Keytool