+++ src/codemagic/tools/android_keystore/__main__.py
@@ -1,0 +1,3 @@
+if __name__ == '__main__':
+    from .android_keystore import AndroidKeystore
+    AndroidKeystore.invoke_cli()