+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+Version 21.0.0
+-------------
+
+This release contains updates from [PR #206](https://github.com/codemagic-ci-cd/cli-tools/pull/206).
+
+**Features**
+- Add new tool `android-keystore` to Android app code signing keystores. New actions are:
+  - `android-keystore create` to initialize new Android keystore,
+  - `android-keystore create-debug-keystore` to initialize new debug Android keystore with default settings,
+  - `android-keystore verify` to check that Android keystore alias and passwords are correct.
+
+**Development**
+- Add optional `env` keyword argument to `CliProcess.execute` method to specify process specific environment variables.
+- Add dataclass `CertificateAttributes` to store certificate issuer information.
+- Add dataclass `Keystore` to store keystore information.
+- Add new module `codemagic.shell_tools` to contain Python wrappers for command line utilities.
+- Add minimal Python wrapper for `keytool` command line utility in `codemagic.shell_tools.Keytool`.
+
+
+**Docs**
+- Add documentation for tool `android-keystore`.
+- Add documentation for action `android-keystore create`. 
+- Add documentation for action `android-keystore create-debug-keystore`. 
+- Add documentation for action `android-keystore verify`. 
+
 Version 20.0.0
 -------------